Extract command validation for easier reuse
